requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
